refactor(api): split authenticate from user service
