+++ packages/api/src/modules/user/__fixtures__/user.fixture.ts
@@ -21,3 +21,5 @@ export const USER_SECRETS = {
 };
 
 export const USER_DBO = { ...USER_WITHOUT_SECRET, ...USER_SECRETS };
+
+export const CONSUMER_USER = { ...USER_WITHOUT_SECRET, roles: ["user", "consumer"] };
