+++ packages/api/tests/authentication/express.auth.hooks.test.ts
@@ -2,10 +2,9 @@ import passport = require("passport");
 import { Express } from "express";
 import passportLocal = require("passport-local");
 import passportJwt = require("passport-jwt");
-import userService from "../modules/user/user.service";
-import { authMocks } from "./express.auth.hooks";
+import { authMocks } from "../../src/authentication/express.auth.hooks";
 import { ObjectId } from "mongodb";
-import userAuthService from "../modules/user/services/auth/user.auth.service";
+import userAuthService from "../../src/modules/user/services/auth/user.auth.service";
 
 describe("express.auth.hooks", () => {
     let passportMock: jest.SpyInstance;
@@ -92,8 +91,8 @@ describe("express.auth.hooks", () => {
             }
 
             jest.spyOn(passportJwt, "Strategy").mockImplementation(strat as any);
-            // @ts-expect-error: mock user
-            jest.spyOn(userService, "authenticate").mockImplementation(async user => ({
+            // @ts-expect-error
+            jest.spyOn(userAuthService, "authenticate").mockImplementation(async user => ({
                 email: user.email,
                 roles: [],
                 active: true,
@@ -129,7 +128,7 @@ describe("express.auth.hooks", () => {
             }
 
             jest.spyOn(passportJwt, "Strategy").mockImplementation(strat as any);
-            jest.spyOn(userService, "authenticate").mockRejectedValue(ERROR);
+            jest.spyOn(userAuthService, "authenticate").mockRejectedValue(ERROR);
 
             passportMock.mockImplementation(name => {
                 if (name === "login") return;
